+++ ReleaseNotes.md
@@ -1,3 +1,4 @@
+* **\[MDAPI-77]\[C++]** Added support for retrieving indexed events from feed
 * **\[MDAPI-36]\[C++]** Implemented DXFeedTimeSeriesSubscription
     * Added `DXFeedTimeSeriesSubscription` class.
     * Added `DXFeed::createTimeSeriesSubscription` methods.